requirements.txt
# Data
yfinance>=0.2.31
pandas>=2.0.0
numpy>=1.24.0

# ML & Statistics
scikit-learn>=1.3.0
statsmodels>=0.14.0
tensorflow>=2.13.0

# Sentiment Analysis
vaderSentiment>=3.3.2
requests>=2.31.0
beautifulsoup4>=4.12.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
